homework/andrii_holovko/requirements.txt
pandas==1.0.5
PyYAML==5.3.1
scikit-learn==0.23.1